+++ core/src/selection/getFragmentNode.ts
@@ -1,6 +1,9 @@
 interface ElementWithDataSet extends HTMLElement {
   readonly dataset: {
-    fragmentStart: string,
+    start: string,
+    end: string,
+    fragment?: string,
+    text?: string,
     blockKey: string
   }
 }
@@ -10,7 +13,7 @@ export default function getFragmentNode (el: HTMLElement | null): ElementWithDat
     return null
   }
 
-  if (el.dataset && el.dataset.blockKey != null && el.dataset.fragmentStart != null) {
+  if (el.dataset && el.dataset.blockKey != null && el.dataset.start != null) {
     const _el: any = el
     return _el
   } else if (el.parentElement) {