+++ react/src/EditorText.tsx
@@ -1,31 +1,33 @@
 import * as React from 'react'
-import { Block, Fragment, TextFragment } from '@zettel/core'
+import { Block, TextOrFragment, Text } from '@zettel/core'
 import { RenderProps } from './types'
 
 type TextProps = RenderProps & {
   block: Block,
 }
 
-const mapTextFramgent = (props: TextProps, offset: number, fragment: TextFragment) => {
+const mapTextFramgent = (props: TextProps, offset: number, fragment: Text) => {
   const {
     block,
     renderStyle: RenderStyle,
-    renderTextFragment: RenderTextFragment,
+    renderText: RenderText,
   } = props
 
   const key = `${block.blockKey}-${offset}`
 
+  const { styles = [], text } = fragment
+
   const fragmentProps = props.readOnly ? {} : {
     key,
     'data-block-key': block.blockKey,
-    'data-text-fragment': true,
-    'data-fragment-start': offset,
-    'data-fragment-end': offset + fragment.text.length
+    'data-text': true,
+    'data-start': offset,
+    'data-end': offset + text.length
   }
 
-  const fragmentText: any = fragment.text
+  const fragmentText: any = text
 
-  let textFragment: React.ReactElement = (fragment.styles || []).reduce((children, val) => {
+  let textFragment: React.ReactElement = styles.reduce((children, val) => {
     if (RenderStyle != null) {
       return <RenderStyle key={`${fragmentProps.key}-${val}`} style={val}>{children}</RenderStyle>
     } else {
@@ -36,36 +38,57 @@ const mapTextFramgent = (props: TextProps, offset: number, fragment: TextFragmen
       key={`fragment-${block.blockKey}-${offset}`}
     >{fragmentText}
   </span>)
-  
 
-  if (RenderTextFragment) {
-    textFragment = <RenderTextFragment
-      data-text-fragment="true"
+  if (RenderText) {
+    textFragment = <RenderText
+      fragment={fragment}
+      data-text="true"
       key={`block-${block.blockKey}-${offset}`}
-    >{textFragment}</RenderTextFragment>
+    >{textFragment}</RenderText>
   }
 
   return textFragment
 }
 
-const reduceFragments = (props: TextProps, _offset: number = 0, fragments: Fragment[]): { rendered: any[], offset: number } => {
+const reduceNodes = (props: TextProps, _offset: number = 0, fragments: TextOrFragment[]): { rendered: any[], offset: number } => {
+  const { block, renderFragment: RenderFragment } = props
+  // @ts-ignore
   return fragments.reduce(
+    // @ts-ignore
     ({ offset, rendered }, fragment) => {
-      if ('fragments' in fragment) {
-        const reducedFragments = reduceFragments(props, offset + 1, fragment.fragments)
-
-        return {
-          rendered: rendered.concat([reducedFragments.rendered]),
-          offset: reducedFragments.offset + 1
-        }
-      } else {
+      if ('text' in fragment) {
         const renderedFragment = mapTextFramgent(props, offset, fragment)
         return {
           offset: offset + fragment.text.length,
           rendered: rendered.concat([renderedFragment])
         }
-      }
+      } else if (RenderFragment != null) {
+        const reducedFragments = reduceNodes(props, offset + 1, fragment.fragments)
+
+        const fragmentProps = props.readOnly ? {} : {
+          key: `block-${block.blockKey}-${offset}-container-fragment`,
+          'data-block-key': block.blockKey,
+          'data-fragment': true,
+          'data-start': offset,
+          'data-end': reducedFragments.offset
+        }
+        
+        const containerFragment = <RenderFragment
+          fragment={fragment}
+          {...fragmentProps}
+        >
+          <>{reducedFragments.rendered}</>
+        </RenderFragment>
+
+        return {
+          rendered: rendered.concat([containerFragment]),
+          offset: reducedFragments.offset + 1
+        }
+      } else {
+        return null
+      } 
     },
+    // @ts-ignore
     { offset: _offset, rendered: ([] as any[]) }
   )
 }
@@ -74,27 +97,29 @@ export default function EditorText(props: TextProps) {
   const {
     block,
     renderStyle: RenderStyle,
-    renderTextFragment: RenderTextFragment,
+    renderText: renderText,
   } = props
 
   let textFragments: React.ReactNode = null
 
-  if (block.value.length > 0) {
+  if (block.value.length === 0 && block.fragments.length === 0) {
+    textFragments = <span
+      key={`text-fragments-${block.blockKey}`}
+      data-text="true"
+      data-block-key={block.blockKey}
+      data-start={0}
+      data-end={0}
+    ><br /></span>
+
+  } else {
     /*
     * If the block has content, split it up into fragments and render the fragments
     */
-    textFragments = reduceFragments(props, 0, block.fragments).rendered
+    // @ts-ignore
+    textFragments = reduceNodes(props, 0, block.fragments).rendered
   /*
    * Render an empty block
    */
-  } else {
-    textFragments = <span
-      key={`text-fragments-${block.blockKey}`}
-      data-text-fragment="true"
-      data-block-key={block.blockKey}
-      data-fragment-start={0}
-      data-fragment-end={0}
-    ><br /></span>
   }
 
   return <>{textFragments}</>