+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Zettel
 
 ## [Unreleased]
+
+### Changed
+- Split out `TextFragment`s into `Text` and `Fragment` to draw clear distinction
+
+### Fixed
+- Rendering container fragments
+
+### Added
 - Replace `ListState` with `Value` type for representing content.
+- Added fragments example
+- Added mentions example
 
 ## [0.0.23]